아이콘 2종 디자인, Kamaitach 잔상 남기기, 핸드폰 샤드 뽑기 트랜지션 더 빠르게 개선
1. 아이콘 2종 디자인
    - Metal Sword 아이콘 디자인
    - Kamaitachi 아이콘 디자인
2. Kamaitach 잔상 남기기
    - 잔상 남기기 프리팹 리뉴얼
    - 기존의 스프라이트 및 애니메이터 방식을 버리고
    - 파티클 시스템으로 스프라이트를 한번 나타냈다가 투명해지며 사라지는 프리팹으로 전환
    - 해당 프리팹을 소환하면 초기화 할때 참조된 SpriteRenderer 혹은 Sprite에서 스프라이트 가져와서 입력하고 1회 나타낸뒤에 디스폰됨
3. 핸드폰 샤드 뽑기 트랜지션 더 빠르게 개선
    - 기존에는 마법 합성 하듯이 등급 가챠를 한번 한뒤에 슬롯을 올려서 등급 업그레이드 가챠를 더 했었는데
    - 수정된 버전에서는 슬롯을 바로 올리면서 등급 노출하고 올려진 뒤에 등급 업그레이드 시도하기

+++ Assets/System/MagicDB.cs
@@ -36,7 +36,7 @@ public class MagicInfo : SlotInfo
     public float criticalPower = 1f; //크리티컬 데미지 증가율
     public int pierce = 0; //관통 횟수 및 넉백 계수
     public int atkNum = 0; //투사체 수
-    public int coolTime = 0; //쿨타임
+    public float coolTime = 0; //쿨타임
 
     [Header("LevUp")]
     public float powerPerLev;
@@ -83,7 +83,7 @@ public class MagicInfo : SlotInfo
     }
 
     public MagicInfo(int id, int grade, string magicName, string element_A, string element_B, string castType, string description, string priceType, bool multiHit, int price,
-    float power, float speed, float range, float duration, float critical, float criticalPower, int pierce, int atkNum, int coolTime,
+    float power, float speed, float range, float duration, float critical, float criticalPower, int pierce, int atkNum, float coolTime,
     float powerPerLev, float speedPerLev, float rangePerLev, float durationPerLev, float criticalPerLev, float criticalPowerPerLev, float piercePerLev, float atkNumPerLev, float coolTimePerLev)
     {
         this.id = id;
